requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
google-generativeai==0.8.3
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0
aiofiles==0.8.0
authlib==1.2.0
itsdangerous==2.1.2
httpx==0.24.1
email-validator==1.3.1
supabase==2.0.0
starlette==0.27.0
moviepy==1.0.3
fastapi-cache2==0.2.1
fastapi-limiter==0.1.5
redis==4.5.5
cachetools==5.3.2
jinja2
pyjwt
ddtrace
ddtrace>=2.0.0
